bike_store/requirements.txt
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.12.1
